Allow update attributes with nil or empty string
Carrierwave will not attempt to set a blank string as valid value, nor
will it delete a file if there is one uploaded. But objects for which
the mounted uploader is nil and we do a mass update_attributes with i.e.
empty string or nil, would result in _old_ being persisted in the attribute.

This fixes that.